--- frontend/src/file/FileProcessor.ts
@@ -1,9 +1,0 @@
-const FileProcessor = (formData: HTMLFormControlsCollection, requestId: string, fileId: string) => {
-    function process() {
-
-    }
-
-    function download() {
-
-    }
-}
