+++ frontend/src/processing-container/filerow/Dummy.tsx
@@ -1,0 +1,6 @@
+type FileSizeText = {
+}
+const FileSizeText = ({}: FileSizeText) => {
+    return <></>;
+}
+export default FileSizeText;
