+++ frontend/src/processing-container/filerow/file.css
@@ -21,7 +21,6 @@
 }
 
 .fade-overflow {
-
     text-overflow: ellipsis;
     white-space: nowrap;
     overflow: hidden;
@@ -30,7 +29,6 @@
 .fs-7 {
     font-size: 0.7rem;
 }
-
 
 .Dropdown-menu {
     width: max-content !important;
@@ -43,4 +41,19 @@
 
 .Dropdown-control {
 
+}
+
+html{
+    background: #eeeeee;
+}
+
+.delete-btn:hover {
+    color: var(--bs-danger);
+}
+
+.add-files-btn{
+    color: var(--bs-dark);
+}
+.add-files-btn:hover{
+    color: var(--bs-black);
 }
