+++ frontend/src/processing-container/utils/FileType.ts
