+++ frontend/src/processing-container/utils/ProcessingStateType.ts
@@ -1,0 +1,1 @@
+type ProcessingStateType = "initial" | "processing" | "processed";
