+++ frontend/src/processing-container/utils/TabType.ts
@@ -1,0 +1,1 @@
+export type TabType = "Convert" | "Compress" | "Merge";
